vold: Only use EXTERNAL_STORAGE for SEC_ASECDIR (cm-issue 2470)
When there are multiple managed volumes, vold was going through the
createBindMounts() dance for /mnt/secure/asec on all of them.  This
made the effective asec mount unpredictable, so apps using that space
would tend to suddenly go missing.

This patch makes it so then environment variable EXTERNAL_STORAGE is the
only volume used for /mnt/secure/asec.

Change-Id: I2f09e594ee25e8efd853775fc0a28de51b4a5db2
Signed-off-by: Josh Stone <cuviper@gmail.com>